FIX Issues and Add Validation

+++ EnergyScore/EnergyScore.Application/Mappers/DTOS/AirInfiltrationMeasurementDTO.cs
@@ -4,7 +4,7 @@ namespace EnergyScore.Application.Mappers.DTOS
     public class AirInfiltrationMeasurementDTO
     {
         public Guid Id { get; set; }
-        public double HousePressure { get; set; }
+        public float HousePressure { get; set; }
         public string UnitofMeasure { get; set; }
         public float AirLeakage { get; set; }
         public string LeakinessDescription { get; set; }
